Implemented Chess\Variant\RandomPosition
