+++ src/Variant/CapablancaFischer/StartPosition.php
@@ -2,56 +2,24 @@
 
 namespace Chess\Variant\CapablancaFischer;
 
+use Chess\Variant\RandomPosition;
 use Chess\Variant\Capablanca\PGN\AN\Piece;
 
-class StartPosition
+class StartPosition extends RandomPosition
 {
-    private $default = [
-        Piece::R,
-        Piece::N,
-        Piece::A,
-        Piece::B,
-        Piece::Q,
-        Piece::K,
-        Piece::B,
-        Piece::C,
-        Piece::N,
-        Piece::R,
-    ];
-
-    public function create()
+    public function __construct()
     {
-        do {
-            shuffle($this->default);
-        } while (!$this->bishops() || !$this->king());
-
-        return $this->default;
-    }
-
-    public function getDefault(): array {
-        return $this->default;
-    }
-
-    private function bishops()
-    {
-        $keys = [];
-
-        foreach ($this->default as $key => $val) {
-            if ($val === Piece::B) {
-                $keys[] = $key;
-            }
-        }
-
-        $even = $keys[0] % 2 === 0 && $keys[1] % 2 === 0;
-        $odd = $keys[0] % 2 !== 0 && $keys[1] % 2 !== 0;
-
-        return !$even && !$odd;
-    }
-
-    private function king()
-    {
-        $str = implode('', $this->default);
-
-        return preg_match('/^(.*)R(.*)K(.*)R(.*)$/', $str);
+        $this->default =  [
+            Piece::R,
+            Piece::N,
+            Piece::A,
+            Piece::B,
+            Piece::Q,
+            Piece::K,
+            Piece::B,
+            Piece::C,
+            Piece::N,
+            Piece::R,
+        ];
     }
 }
